--- Source/Constants.swift
@@ -1,11 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Lukas Schmidt on 17.04.20.
-//
-
-import Foundation
-
-let ROOT_ID: String = "00000000-0000-0000-0000-000000000000"
-
