+++ Source/DataTypes/Object.swift
@@ -7,31 +7,6 @@
 
 import Foundation
 
-public struct ObjectId: Equatable, Hashable, Codable, ExpressibleByStringLiteral {
-
-    init(objectId: String = UUID().uuidString) {
-        self.objectId = objectId
-    }
-
-    let objectId: String
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        self.objectId = try container.decode(String.self)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(objectId)
-    }
-
-    static let root = ObjectId(objectId: ROOT_ID)
-
-    public init(stringLiteral value: StringLiteralType) {
-        self.objectId = value
-    }
-}
-
 enum Object: Equatable, ExpressibleByStringLiteral, ExpressibleByFloatLiteral, ExpressibleByArrayLiteral, Codable {
 
     case text(Text)
